split the meshlab script
Split into solidification and simplification scripts.
We can work around a bug in meshlabserver that crashes
on simplification by first making the solid and saving it
as a .obj, and then running simplification on that and
exporting to an .stl.  This allows for the possibility of
running meshlabserver from the python script.